[Gold III] Title: 나머지 합, Time: 496 ms, Memory: 116636 KB -BaekjoonHub

+++ 백준/Gold/10986. 나머지 합/나머지 합.java
@@ -11,18 +11,18 @@ public class Main {
         int m = Integer.parseInt(st.nextToken());
 
         long cnt = 0;
-        long[] sum = new long[n + 1];
-        long[] remainder = new long[m];
+        int[] sum = new int[n + 1];
+        int[] remainder = new int[m];
         st = new StringTokenizer(br.readLine());
         for (int i = 1; i <= n; i++) {
-            long num = Integer.parseInt(st.nextToken());
+            int num = Integer.parseInt(st.nextToken());
             sum[i] = (sum[i - 1] + num) % m;
             if (sum[i] == 0) cnt++;
-            remainder[(int) sum[i]]++;
+            remainder[sum[i]]++;
         }
 
         for (int i = 0; i < m; i++) {
-            cnt += remainder[i] * (remainder[i] - 1) / 2;
+            cnt += (long) remainder[i] * (remainder[i] - 1) / 2;
         }
 
         System.out.println(cnt);
